requirements-dev.txt
factory-boy==3.2.1
pip-tools
pre-commit
pytest-django
pytest-xdist
pytest-cov==4.1.0 
pytest-socket==0.6.0 
pygraphviz==1.11